+++ crates/assembly-syntax/src/ast/op.rs
@@ -2,7 +2,7 @@ use core::fmt;
 
 use miden_debug_types::{SourceSpan, Span, Spanned};
 
-use super::{Block, Instruction};
+use super::{Block, Immediate, Instruction};
 
 /// Represents the Miden Assembly instruction set syntax
 ///
@@ -27,9 +27,10 @@ pub enum Op {
     ///
     /// NOTE: The iteration count must be known at compile-time, so this is _not_ used for general
     /// `for`-style loops where the iteration count is dynamic.
+    /// Also, it may be either a literal value or constant name (e.g. `add.1` or `add.CONST`)
     Repeat {
         span: SourceSpan,
-        count: u32,
+        count: Immediate<u32>,
         body: Block,
     } = 2,
     /// A primitive operation, e.g. `add`